requirements.txt
gunicorn==20.1.0
dj-database-url==0.5.0
whitenoise==5.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
anytree==2.8.0
asgiref==3.5.2
certifi==2021.10.8
charset-normalizer==2.0.7
channels==4.0.0
cloudinary==1.26.0
crispy-bootstrap4==2022.1
dj-database-url==0.5.0
Django>=3.2.7
django-cleanup==5.2.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.3
Pillow==8.3.2
pip==22.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2021.1
requests==2.26.0
setuptools==57.2.0
six==1.16.0
sqlparse==0.4.1
tzdata==2022.2
urllib3==1.26.7
wheel==0.36.2
whitenoise==5.3.0
xlwt==1.3.0